+++ src/main/server/skyRepo.js
@@ -103,6 +103,7 @@ const getTypeFromObject = function (o) {
         return context + '/' + type;
     }
 };
+module.exports.getTypeFromObject = getTypeFromObject;
 
 const signatureSheet = async function () {
     let sigSheet = null;
@@ -274,6 +275,7 @@ const filterResults = async function (o, dontDecryptInSso) {
 const skyrepoUrlType = function (o) {
     return getTypeFromObject(o);
 };
+module.exports.skyrepoUrlType = skyrepoUrlType;
 const inferTypeFromObj = function (o, atType) {
     // if (atType != null)
     //     return atType;
@@ -283,14 +285,11 @@ const inferTypeFromObj = function (o, atType) {
     }
     fullType = fullType.replace('http://', '');
     fullType = fullType.replace('https://', '');
-    fullType = fullType.replace('/', '.');
-    fullType = fullType.replace('/', '.');
-    fullType = fullType.replace('/', '.');
-    fullType = fullType.replace(':', '.');
-    fullType = fullType.replace(':', '.');
-    fullType = fullType.replace(':', '.');
+    fullType = fullType.replace(/\//g, '.');
+    fullType = fullType.replace(/:/g, '.');
     return fullType;
 };
+module.exports.inferTypeFromObj = inferTypeFromObj;
 const inferTypeWithoutObj = function (atType) {
     if (atType !== undefined && atType != null) {
         return atType;